Add skeletal span support.
This adds a structure `Spanned<T>` and wraps most AST nodes in it.
Currently the `Span` structure is unpopulated, but should in the future
contain actual source locations that can be used for reporting errors.

Parser support for spans is planned, in the form of a `span!()` macro
that will expand to the span of the current rule's RHS. This will
probably use some kind of `HasSpan` trait, which will allow the parser
to extract spans from tokens. The parser will probably keep track of
spans internally.

Identifiers also now have their own spans, i.e. Ident = Spanned<String>.
I'm not 100% convinced this makes sense though - does the compiler ever
need to generate its own Idents?

To make pattern matching a bit nicer, a `node!()` macro is introduced.