Get rid of `Name`s, replacing them with pointers.
This removes the extra level of indirection through the hash table, and
hopefully makes the meaning more clear. It also adds some type-safety,
since we can hold references to specific types, instead of having to go
through NamedItem every time. However, it adds a ton of lifetime noise
(<'all ,'ast>) and might not be as easy to work with.

Alternatives:
- Use Name, but use some type-safe abstraction on top of it (Name<T>
  with a map from Name<T> to T).
- Just go with what we have today.

Thoughts?