Arenas, take 2.
This gets rid of Rc for the object graph, in favour of an arena. This
removes the difference between strong and weak references, and removes a
lot of `.clone`s, in exchange for some more noise around function
signatures.

Also, the manual impl of `Eq` and `Ord` for `SimpleType` is replaced
with an auto-derived one.