Implement runtime type checks for interfaces.
This should allow arrays of interface types to interact correctly with
casts, `instanceof`, and index assignment.

This uses a bit of a trick to distinguish interface descriptors from
type descriptors: type descriptors are aligned on an 8-byte boundary
while interface descriptors are off-aligned by 4 bytes. To test this,
AND with 0x04.